trascan-be/requirements.txt
fastapi
uvicorn
tensorflow
google-cloud-storage
numpy
pillow
python-multipart
google-cloud-firestore
